Copper Wire Connection Solution
You can use a hammer and a stithy to nail the copper wire to a thin
piece and then use a scissor to modify it to a proper shape.

Finally, insert the copper piece into the leak between the framework
and the rail.

Attention: Please make sure to install the copper wire between the
outer surface of the rail and the outer side of the framework!